dbeaver/pro#6823 Add a short error message for Gemini
